split-out summary stats from tile stats
